+++ src/py/flwr/client/app_test.py
@@ -17,6 +17,8 @@
 
 from typing import Dict, Tuple
 
+from flwr.client.numpy_client_wrapper import to_client
+from flwr.client.typing import ClientLike
 from flwr.common import (
     Config,
     EvaluateIns,
@@ -31,7 +33,7 @@ from flwr.common import (
     Scalar,
 )
 
-from .app import ClientLike, start_client, start_numpy_client, to_client
+from .app import start_client, start_numpy_client
 from .client import Client
 from .numpy_client import NumPyClient
